move a bunch into helpers
